Added Loops and ++ and -- Operators
Added the forLoop grammar and function to handle each part of a for loop.
Added the while loop grammar and function to handle a boolean loop.

Fixed how ifelse statements use statement and updated them to work with the change.

Added a for loop and while loop to the input.dcf file.